requirements.txt
pyqt6
pysdl2
